Fix iOS builds from command line (xcodebuild)
For whatever reason, it seems when building inside Xcode.app, the
amalgamation target has environment variables set as if it is being
built for macOS (even when building the iOS library), so the
configure/make steps work fine.

However, when building via the command line (ie: `xcodebuild -scheme
sqlitelib -configuration Release -sdk iphoneos`), the amalgamation
target was getting environment variables set as if it was building for
iOS. This was causing the `configure` step to fail, because it thought
the compiler was broken (macOS couldn't run binaries that were being
built for iOS).

This commit fixes the issue by always running configure as if we were
building for macOS. (The sqlitecustom target will still build libraries
for the correct platform, as it did before)