make test
Now that Travis is gone, we need some kind of automated check.

Until someone defines, for example, a GitHub action, this commit adds a sanity check that we can build a SQLite library and link with it, with `make test`.

Also, add `make distclean` for restoring the repository in a pristine state.